Moved UIShell to content
